unified files
loginResponse-userResponse and loginRequest-userRequest services are unified.
