+++ app/src/main/java/com/homies/homies/LoginActivity.java
@@ -4,6 +4,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.os.Parcelable;
 import android.text.TextUtils;
 import android.view.View;
 import android.widget.Button;
@@ -48,24 +49,24 @@ public class LoginActivity extends AppCompatActivity {
                     String message = "Password must have between 8 and 100 characters ...";
                     Toast.makeText(LoginActivity.this, message, Toast.LENGTH_SHORT).show();
                 } else {
-                    LoginRequest loginRequest = new LoginRequest();
-                    loginRequest.setUsername(inputUser.getText().toString());
-                    loginRequest.setPassword(inputPassword.getText().toString());
+                    UserRequest userRequest = new UserRequest();
+                    userRequest.setUsername(inputUser.getText().toString());
+                    userRequest.setPassword(inputPassword.getText().toString());
 
-                    loginUser(loginRequest);
+                    loginUser(userRequest);
                 }
             }
         });
     }
 
-    public void loginUser(LoginRequest loginRequest) {
-        Call<LoginResponse> loginResponseCall = ApiClient.getService().loginUser(loginRequest);
-        loginResponseCall.enqueue(new Callback<LoginResponse>() {
+    public void loginUser(UserRequest userRequest) {
+        Call<UserResponse> loginResponseCall = ApiClient.getService().loginUser(userRequest);
+        loginResponseCall.enqueue(new Callback<UserResponse>() {
             @Override
-            public void onResponse(Call<LoginResponse> call, Response<LoginResponse> response) {
+            public void onResponse(Call<UserResponse> call, Response<UserResponse> response) {
                 if (response.isSuccessful()) {
-                    LoginResponse loginResponse = response.body();
-                    startActivity(new Intent(LoginActivity.this, MainActivity.class).putExtra("data", loginResponse));
+                    UserResponse userResponse = response.body();
+                    startActivity(new Intent(LoginActivity.this, MainActivity.class).putExtra("data", (Parcelable) userResponse));
                     finish();
                 } else {
                     String message = "An error ocurred please try again later ...";
@@ -74,7 +75,7 @@ public class LoginActivity extends AppCompatActivity {
             }
 
             @Override
-            public void onFailure(Call<LoginResponse> call, Throwable t) {
+            public void onFailure(Call<UserResponse> call, Throwable t) {
                 String message = t.getLocalizedMessage();
                 Toast.makeText(LoginActivity.this, message, Toast.LENGTH_SHORT).show();
             }
