--- app/src/main/java/com/homies/homies/LoginResponse.java
@@ -1,34 +1,0 @@
-package com.homies.homies;
-
-import java.io.Serializable;
-
-public class LoginResponse implements Serializable {
-
-    private String username;
-    private String password;
-    private String token;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-}
