+++ app/src/main/java/com/homies/homies/UserRequest.java
@@ -7,6 +7,8 @@ public class UserRequest {
     private String firstName;
     private String lastName;
     private String login;
+    private String username;
+    private String token;
 
     public String getPassword() {
         return password;
@@ -47,4 +49,22 @@ public class UserRequest {
     public void setLogin(String login) {
         this.login = login;
     }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
 }
